# Add generic fast format converters for better performance