--- a/pyproject.toml
+++ b/pyproject.toml
@@ -18,3 +18,10 @@ exclude = '''
 
 [tool.isort]
 profile = "black"
+
+[tool.flake8]
+exclude = '''
+/(
+    docs
+)/
+'''