## Description

Creating a `CxoTime` object is unfortunately slow, taking a few hundred microsecs. Often the only reason for this it so convert to a different format like `date` or `secs`. This PR includes functionality to do those conversions in tens of microsecs (for the scalar case).

Features:
* Includes a functional interface `convert_time_format` that converts between `date`, `secs`, `jd`, `maude` and `greta` formats without creating a `CxoTime` object. It can also infer the input time format in a performant way for common formats (in particular `date` and `secs`).
* Generalize the convenience methods `date2secs` and `secs2date` fast converters to include all combinations of the supported fast formats in `convert_time_format`.
* Update the `CxoTime` initializers so that `CxoTime(tm)` returns the `tm` object if it is already a `CxoTime` instance. This allows writing code like `start = CxoTime(start)` with no performance penalty if `start` is already a `CxoTime`.  `CxoTime(tm, copy=True)` will make a copy as expected.'

This is based on the `remove-cxotime-fast-parser` branch since that was needed. The intent is to rebase on master once #34  is merged.

### Todo
- [x] Docs
- [x] Add unit tests
- [x] Ensure reasonable docstrings for the convenience conversion functions
- [x] Some code clean-up

## Interface impacts
<!-- API changes, file format updates, coordination of changes with the community. -->
The `date2secs` and `secs2date` functions were removed from the `cxotime.cxotime` so `from cxotime.cxotime import date2secs` will now fail. This was not the public interface and code should be using `from cxotime import date2secs`.

## Testing
<!-- If relevant describe any special setup for testing. -->

### Unit tests
<!-- At least one of these must be checked if unit tests exist. DELETE the unchecked/untested options. -->
- [x] Mac


Independent check of unit tests by Jean
- [x] Linux (in a dev ska3-matlab-2023.4rc6 env)

### Functional tests
<!-- Describe and document results of any functional tests, otherwise leave the text below -->
Roughly speaking using `CxoTime` for a single value is around 350 µs, while the "fast" functions will take around 25 µs.
```
>>> from cxotime import *
>>> %timeit CxoTime(100.0).greta
373 µs ± 6.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> %timeit secs2greta(100.0)
24.4 µs ± 577 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit convert_time_format(100.0, 'greta')
28.1 µs ± 1.71 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit convert_time_format(100.0, 'greta', fmt_in='secs')
24.2 µs ± 457 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit greta2secs(2020001.123456789)
25.4 µs ± 1.85 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit greta2secs('2020001.123456789')
20.7 µs ± 1.61 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
Ultimate conversion speed is about 0.2 µs / conversion for large arrays:
```
>>> x = ["2020001.123456789"] * 1000
>>> %timeit greta2secs(x)
358 µs ± 5.43 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> x = np.array(["2020001.123456789"] * 1000)
>>> %timeit greta2secs(x)
218 µs ± 4.21 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```