@taldcroft 2023-07-06
I rebased this and tests are passing locally on Mac. All good? We'll see in integration testing if something comes up?

I updated the Interface section to note the removal of `cxotime.cxotime.date2secs/secs2date`.